Email notification on project registration/cancellation (#781)
Code for sending email confirmations once a user has registered or de-registered for a project. Upon registering, the an email is dispatched, containing an `.ics` calendar file and details about the project, including project title, information about the next session in the schedule, and project URL. Upon de-registering for a project an email is dispatched to the user confirming the same.