Remove strip filter for UrlType fields (#1158)
Strip filters were added for all text and URL fields in #1156. However, `UrlType` columns return a `Furl` type, which casts to a string but has no `strip` method, causing an internal server error when the form is processed.

Ideally the strip filter should accept a `coerce=str` parameter, so that it can be processed as a string before being converted back into a `Furl`.