Fix and test SMS unsubscribe view with new form submit wrapper (#1216)
The SMS unsubscribe view was broken by #1207 due to a mistaken assumption about datetimes. While the Flask 2.0 `session` object now stores tz-aware datetimes, Flask-Caching does not do this. Datetimes stored in Redis continue to be naive.

This PR fixes the bug and also introduces tests for the view, to confirm the SMS unsubscribe form is accessible and can be submitted successfully.

A new `ResponseWithForms` test helper class is introduced for the test client fixture, making it possible to test forms using the actually rendered HTML.