Allow login with a phone number (#1326)
"Username or email" has changed to "Email, phone or username", and numbers are assumed to be Indian or US if not in international format.

This is a partial implementation towards #1097. The next steps are to (a) allow login with an OTP, (b) allow account creation with a phone number, and (c) embed a login or register form within the project registration prompt.