Fix reload_for_cookies decorator
The decorator was suppressing the setting of a cookie if the request began without one. However, the login callback decorator is an instance where it is legitimate to start with no incoming cookie. This situation was missed earlier because testing started from the login page, which does set a cookie. Starting from a login service redirect exposed the bug.