Test imports cleanup (#1467)
This PR started as a failed attempt to remove all non-stdlib imports from tests, moving them into fixtures. This would enable pylint with the pylint-pytest plugin to run in pre-commit, where it has an isolated venv without the dependencies necessary for import. Pylint can be asked to ignore missing imports, but pylint-pytest asks pytest to collect fixtures, and that fails when imports are missing.

This could not be completed as imports are used for parametrizing tests in `tests/unit/views/test_search.py`, and moving parametrization into a fixture was making it too complicated. Dependency injection also breaks IDE auto-completion and mypy type checking, making test development that much harder. Therefore we continue with status quo of not running pylint on the tests folder in pre-commit.

However, some of the re-organization remains. Imports are now mostly namespaced (`models` and `forms`) and spurious warnings from `failsafe_add` are now suppressed.