Fix Babel processing in two places
Lazy strings are required to be HTML-escaped at source and have a `__html__` attr that marks them as safe, but turns out `lazystr.format()` produces a plain `str` and has to be recast as Markup again.

Flask-Babel has wrappers for Babel functions that automatically supply locale, so this commit switches over.