Change cookie serializer (#1710)
Funnel's cookie serializer used JWT, which was removed from itsdangerous, forcing us to pin the old version. We could not switch to another JWT implementation because our key rotation wrapper only works with itsdangerous. Rather than extend key rotation to another library, with fragile tests, we've opted to switch to a supported serializer in itsdangerous. This switch will invalidate all existing cookies, forcing a re-login for all users. A corresponding change in hasgeek/flask-lastuser#66 is required for all apps hosted in subdomains of Funnel.

In addition, RQ Dashboard now requires config on register and so has a new init function. Flask 2.3 no longer supports the `before_first_request` callback, so the blueprint must expect config from the app when first registered on it.