Grant project_* roles within Project too, and drop transition validator (#1716)
ProjectStartingNotification depends on the `project_crew` and `project_participant` roles, which it expects to find in the Session model, but which are not granted in the project itself. The notification is also dispatched on projects with no session, causing a discovery of 0 recipients.

The `if_` validator on the `Profile.make_public` transition is redundant as a conditional state serves the same purpose. Coaster will be removing transition validators.