Gracefully handle email recipient rejection by the SMTP server (#1748)
A string in the form `anything@123` will pass as an email address in the login form, but will be rejected by the SMTP server. This patch recasts the SMTP error as a form validation error.

A test is included. Since Flask-Mailman will silently drop email in a test environment, the test has to mock Flask-Mailman instead of the underlying `smtplib.SMTP.sendmail` where the error is raised in production.