Send Updates to participants across the account (#1763)
The "follower" model on Profile (soon Account) is still pending, but we fake it for now by recasting the "participant" role in any project into a "follower" role on the account. This is an inefficient database lookup but tolerable when processing only a few thousand followers. How it works:

1. Project currently grants a `participant` role via the Rsvp and ProjectCrewMembership models
2. Profile now grants `follower` role for any actor who has a `participant` role in a listed project
3. Project in turn recasts `follower` from Profile as `account_participant`
4. Update also acquires `account_participant` from Project, but only when the Update is public
5. NewUpdateNotification queries Update for all actors with `account_participant` role, which Update now defines as everyone in the account, or nobody when the update is restricted

Since role grants are lazy lookups, this recursive definition has no cost until when the `account_participant` role is actually needed, which is currently only in the notification dispatcher that runs as a background process. Regular interactive use of the website will not be impacted.