Grant ticket_participant role only via an event
The `ticket_participant` role used to be granted to every TicketParticipant instance in a project, but these are synced from a Boxoffice menu that may list multiple unrelated events together. This is a problem for audience-segregation via projects. This commit now grants the role only if the project has an event that the ticket type is assigned to, thereby excluding synced participants who aren't the audience for this project.