Grant ticket_participant role only via an event (#2187)
The `ticket_participant` role used to be granted to every TicketParticipant instance in a project, but these are synced from a Boxoffice menu that may list multiple unrelated events together. This is a problem for audience segregation via Projects. This commit now grants the role only if the Project has a TicketEvent that the TicketType is assigned to, thereby excluding synced TicketParticipants who aren't the audience for this project.